backend/pyproject.toml
[tool.poetry]
name = "test-elefanto"
version = "0.1.0"
description = ""
authors = ["Daniyar <srxd1710@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
punq = "^0.7.0"
djangorestframework = "^3.15.2"
djangorestframework-dataclasses = "^1.3.1"
pytest = "^8.2.2"
pre-commit = "^3.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
line_length = 120
lines_after_imports = 2
lines_between_sections = 1
lines_between_types = 0
force_alphabetical_sort_within_sections = true
ensure_newline_before_comments = true
combine_as_imports = true
combine_star = true
skip_gitignore = true
balanced_wrapping = true
force_grid_wrap = 2
src_paths = ["core", ]
skip_glob = [
    ".git",
    ".idea",
    ".vscode",
    "venv",
    ".venv",
]
known_django = ["django", "ninja"]
known_first_party = ["core"]
sections = [
    "FUTURE",
    "STDLIB",
    "DJANGO",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
